Generate or Load Cached Gamma Predictions

This function generates gamma parameter predictions using spatial model parameters and predictors data, or loads cached predictions if the file already exists. It saves the predictions to a specified directory and provides progress updates.

Usage

generate_gamma_predictions(
  country_code,
  age_param_data,
  model_params,
  predictor_data,
  shapefile,
  cell_size = 5000,
  n_sim = 10000,
  ignore_cache = FALSE,
  save_file = FALSE,
  output_dir = here::here("03_outputs", "3a_model_outputs")
)

Arguments

country_code

A string representing the country code (e.g., "KEN").

age_param_data

A data frame or tibble containing the age parameter data.

model_params

The fitted spatial model object (e.g., `spat_model_param`).

predictor_data

A data object containing the predictors data.

shapefile

An `sf` object representing the country shape.

cell_size

An integer specifying the cell size for prediction in meters (default is 5000).

n_sim

An integer specifying the number of simulations for predictions (default is 10000).

ignore_cache

A boolean input which is set to determine whether to ignore the existing cache and write over it. Default is set to FALSE.

save_file

A boolean to determine whetehr to save prediction or not. Default is FALSE as this will requre lots of space.

output_dir

A string specifying the directory where the predictions file should be saved (default is "03_outputs/3a_model_outputs").

Value

A data object (`gamma_prediction`) containing the gamma parameter predictions.

Examples

# gamma_results <- generate_gamma_predictions(
#  country_code = "KEN",
#  age_param_data = age_param_data,
#  model_params = spat_model_param,
#  predictor_data = predictor_data,
#  shapefile = country_sf,
#  cell_size = 5000,
#  n_sim = 10000,
# output_dir = "03_outputs/3a_model_outputs"
#  )