Generate or Load Cached Predictors Data

This function creates predictors data based on spatial inputs or loads cached predictors data if the file already exists. It saves the generated data to a specified directory for reuse and provides progress updates.

Usage

create_prediction_data(
  country_code,
  country_shape,
  pop_raster,
  ur_raster,
  adm2_shape,
  cell_size = 5000,
  ignore_cache = FALSE,
  output_dir = here::here("03_outputs", "3a_model_outputs")
)

Arguments

country_code

A string representing the country code (e.g., "KEN").

country_shape

An `sf` object representing the country's administrative boundaries.

pop_raster

A `terra` raster object representing the population raster.

ur_raster

A `terra` raster object representing the urban extent raster.

adm2_shape

An `sf` object representing the administrative level 2 boundaries.

cell_size

An integer specifying the cell size for the prediction grid in meters (default is 5000).

ignore_cache

A boolean input which is set to determine whether to ignore the existing cache and write over it. Default is set to FALSE.

output_dir

A string specifying the directory where the predictors data file should be saved (default is "03_outputs/3a_model_outputs").

Value

A data object (`predictor_data`) containing the generated predictors.

Examples

if (FALSE) { # \dontrun{
predictors <- create_prediction_data(
  country_code = "KEN",
  country_shape = country_sf,
  pop_raster = pop_raster,
  ur_raster = urban_raster,
  adm2_shape = adm2_sf,
  cell_size = 5000,
  output_dir = "03_outputs/3a_model_outputs"
)
} # }