Main Function to Download DHS Datasets

Downloads the latest PR (household) and GE (geographic) DHS datasets for specified countries.

Usage

download_dhs_datasets(
  country_codes,
  cache_path = here::here("01_data", "1a_survey_data", "raw"),
  output_dir_root = here::here("01_data", "1a_survey_data", "raw"),
  email,
  project,
  verbose = TRUE
)

Arguments

country_codes

A character vector of ISO3 country codes.

cache_path

A character string specifying the cache path for RDHS.

output_dir_root

A character string specifying the root directory for output.

email

A character string. Email registered with DHS.

project

A character string. Project name as registered with DHS.

verbose

Logical for rdhs setup and messages to be printed and dataset download progress bars to be shown.

Value

Invisibly returns a list of downloaded dataset filenames.